# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сезоны — API дашборда",
    "description": "Аналитика продаж ресторана из iiko.\n\nКонтракты: бэкенд отдаёт факты, представление — зона фронтенда. docs/frontend-handoff.md",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/targets": {
      "get": {
        "tags": [
          "Цели"
        ],
        "summary": "Цели: месяц (query) или список ?status=locked|configured",
        "operationId": "listOrGetTargets",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TargetsData"
                    },
                    {
                      "$ref": "#/components/schemas/TargetsLockedList"
                    }
                  ],
                  "title": "Response Listorgettargets"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/targets/{year}/{month}": {
      "get": {
        "tags": [
          "Цели"
        ],
        "summary": "Цели на месяц",
        "operationId": "getTargetsMonth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Цели"
        ],
        "summary": "Сохранить цели месяца",
        "operationId": "putTargetsMonth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Цели"
        ],
        "summary": "Delete Targets Month",
        "operationId": "deleteTargetsMonth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/targets/{year}/{month}/lock": {
      "post": {
        "tags": [
          "Цели"
        ],
        "summary": "Post Lock Month",
        "operationId": "lockTargetsMonth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Цели"
        ],
        "summary": "Снять блокировку (DELETE lock)",
        "operationId": "unlockTargetsMonth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TargetsAmountGoalUpsert": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "pct",
              "rub"
            ],
            "title": "Mode"
          },
          "pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pct"
          },
          "rub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rub"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "pct",
          "rub"
        ],
        "title": "TargetsAmountGoalUpsert",
        "description": "Цель в % или в валюте — как у списаний."
      },
      "TargetsCompliments": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "pct",
              "rub"
            ],
            "title": "Mode",
            "default": "pct"
          },
          "goalPct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Goalpct"
          },
          "goalRub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Goalrub"
          },
          "factPct": {
            "type": "number",
            "minimum": 0.0,
            "title": "Factpct"
          },
          "factRub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Factrub"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goalPct",
          "goalRub",
          "factPct",
          "factRub"
        ],
        "title": "TargetsCompliments"
      },
      "TargetsData": {
        "properties": {
          "period": {
            "$ref": "#/components/schemas/TargetsPeriod"
          },
          "reference": {
            "$ref": "#/components/schemas/TargetsReference"
          },
          "revenue": {
            "$ref": "#/components/schemas/TargetsRevenue"
          },
          "dailyOverrides": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Dailyoverrides",
            "description": "Дневные override: ключ — день месяца (\"1\"…\"31\")"
          },
          "foodcost": {
            "items": {
              "$ref": "#/components/schemas/TargetsFoodcostUnit"
            },
            "type": "array",
            "title": "Foodcost"
          },
          "writeoffs": {
            "items": {
              "$ref": "#/components/schemas/TargetsWriteoffUnit"
            },
            "type": "array",
            "title": "Writeoffs"
          },
          "compliments": {
            "$ref": "#/components/schemas/TargetsCompliments"
          },
          "inventory": {
            "$ref": "#/components/schemas/TargetsInventory"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "reference",
          "revenue",
          "foodcost",
          "writeoffs",
          "compliments",
          "inventory"
        ],
        "title": "TargetsData"
      },
      "TargetsFoodcostUnit": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "goalPct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Goalpct"
          },
          "factPct": {
            "type": "number",
            "minimum": 0.0,
            "title": "Factpct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name",
          "goalPct",
          "factPct"
        ],
        "title": "TargetsFoodcostUnit",
        "description": "GET: goalPct может быть 0 до настройки."
      },
      "TargetsFoodcostUnitUpsert": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "goalPct": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Goalpct"
          },
          "factPct": {
            "type": "number",
            "minimum": 0.0,
            "title": "Factpct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name",
          "goalPct",
          "factPct"
        ],
        "title": "TargetsFoodcostUnitUpsert"
      },
      "TargetsInventory": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "pct",
              "rub"
            ],
            "title": "Mode",
            "default": "pct"
          },
          "goalPct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Goalpct"
          },
          "goalRub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Goalrub"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goalPct",
          "goalRub",
          "note"
        ],
        "title": "TargetsInventory"
      },
      "TargetsLockedList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TargetsLockedPeriod"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TargetsLockedList"
      },
      "TargetsLockedPeriod": {
        "properties": {
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "month",
          "label"
        ],
        "title": "TargetsLockedPeriod"
      },
      "TargetsPeriod": {
        "properties": {
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "month",
          "label"
        ],
        "title": "TargetsPeriod"
      },
      "TargetsReference": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "revenueFact": {
            "type": "number",
            "title": "Revenuefact"
          },
          "revenuePace": {
            "type": "number",
            "title": "Revenuepace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "revenueFact",
          "revenuePace"
        ],
        "title": "TargetsReference",
        "description": "Подсказки «факт / темп» — read-only, считаются с бэкенда."
      },
      "TargetsRevenue": {
        "properties": {
          "monthPlan": {
            "type": "number",
            "minimum": 0.0,
            "title": "Monthplan"
          },
          "weekProfile": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 7,
            "title": "Weekprofile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "monthPlan",
          "weekProfile"
        ],
        "title": "TargetsRevenue",
        "description": "GET: план может быть 0, если месяц ещё не настроен."
      },
      "TargetsRevenueUpsert": {
        "properties": {
          "monthPlan": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Monthplan"
          },
          "weekProfile": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 7,
            "title": "Weekprofile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "monthPlan",
          "weekProfile"
        ],
        "title": "TargetsRevenueUpsert",
        "description": "PUT: все поля обязательны и > 0."
      },
      "TargetsUpsertRequest": {
        "properties": {
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "revenue": {
            "$ref": "#/components/schemas/TargetsRevenueUpsert"
          },
          "dailyOverrides": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Dailyoverrides"
          },
          "foodcost": {
            "items": {
              "$ref": "#/components/schemas/TargetsFoodcostUnitUpsert"
            },
            "type": "array",
            "minItems": 1,
            "title": "Foodcost"
          },
          "writeoffs": {
            "items": {
              "$ref": "#/components/schemas/TargetsWriteoffUnitUpsert"
            },
            "type": "array",
            "minItems": 1,
            "title": "Writeoffs"
          },
          "compliments": {
            "$ref": "#/components/schemas/TargetsAmountGoalUpsert"
          },
          "inventory": {
            "$ref": "#/components/schemas/TargetsAmountGoalUpsert"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "month",
          "revenue",
          "foodcost",
          "writeoffs",
          "compliments",
          "inventory"
        ],
        "title": "TargetsUpsertRequest",
        "description": "Тело PUT — все обязательные поля заполнены (> 0)."
      },
      "TargetsWriteoffUnit": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "pct",
              "rub"
            ],
            "title": "Mode"
          },
          "pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pct"
          },
          "rub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rub"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name",
          "mode",
          "pct",
          "rub"
        ],
        "title": "TargetsWriteoffUnit",
        "description": "GET: суммы могут быть 0 до настройки."
      },
      "TargetsWriteoffUnitUpsert": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w",
              "o"
            ],
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "pct",
              "rub"
            ],
            "title": "Mode"
          },
          "pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pct"
          },
          "rub": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rub"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name",
          "mode",
          "pct",
          "rub"
        ],
        "title": "TargetsWriteoffUnitUpsert"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}